requirements.txt
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
databases==0.5.5
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.73.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
pydantic==1.9.0
python-dotenv==0.19.2
requests==2.27.1
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
toolz==0.11.2
typing-extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.1
